SILOptimizer: Add support for generic closures to CapturePromotion
Previously this pass would bail out if the closure had any
parameters or results of generic type.

First of all, this was a workaround for an old SILGen limitation,
where closures in generic context would inherit a generic
signature even if they did not capture generic types, which
inhibited optimizations.

Second, capture promotion doesn't really do anything special with
generics at all, but it used a TypeSubstCloner and would apply
substitutions from the call site to the cloned function; this is
not the behavior we want here. Simply switching the pass to use a
SILCloner instead of a TypeSubstCloner appears to be sufficient
in order to allow generic boxes to be promoted too.

Fixes <rdar://problem/28948735>.