Introduce special decl names
Special DeclNames represent names that do not have an identifier in the
surface language. This implies serializing the information about whether
a name is special together with its identifier (if it is not special)
in both the module file and the swift lookup table.

ValueDecls are not guaranteed to have a name anymore and thus getNameStr
is moved from ValueDecl to subclasses that guarantee to be backed by
identifiers. Other calls need to handle special names if the context
doesn't guarantee that the declaration's name is non-special.