Preparations for special names
Special names will in the future represent names that don't have a
surface language identifier like subscript.

For this, change the backing type of DeclName to a new type DeclBaseName
that will later be able to represent special names.

ValueDecls may therefore no longer have an identifier name. Thus we move
the getName method from ValueDecl to the subtypes that are guaranteed
to have identifier names