[stdlib] String prototype: bring String size down to 64 bits
This will be required if we're to get COW uniqueness checking working, because
you can't query the enum without bumping the refcount.  We need to ensure that
the enum's address can be passed directly to _isUnique.  Note: we may need to do
more to support 32-bit architectures.  Depends on how things get laid out.

To shrink the storage, flags got moved into StringStorage.  Along the way,
changed the count/capacity storage for the default string to 32-bits, to save
space.  We can always add a storage format that uses 64-bit count/capacity.  We
should think about whether a version with 8-bit count/capacity would be a good
idea.