AST: Add optional GenericEnvironment parameter to TypeBase::getContextSubstitutionMap()
If a generic type appears in a generic function, the old
gatherAllSubstitutions() method would map outer generic
parameters to archetypes.

However, getContextSubstitutionMap() did not map them to
anything.

I'm trying to unify these methods, so add an optional
GenericEnvironment to getContextSubstitutionMap() to support
the cases where they're needed.

Of course types in generic functions are not supported right
now, but not preserving this subtle behavioral difference
makes some crashers regress.