stdlib: Improve String comparison performance
This patch improves String comparison performance by
introducing a memcmp-like helper routine that will perform
a binary comparison on two buffers of equal length N and
return the offset of the first difference if any, or N if
they contain no differences.

The helper is invoked from String operators == (for
non-Objective-C runtimes only) and < (for all runtimes)
when possible, the critera being that the Strings must have
the same element size and must use contiguous storage. If
we detect a mismatch in the binary representation of the
Strings we fall back to using the ICU routines to determine
the result of the comparison. Strings of differing binary
length are handled by invoking the helper to check up to
the length of the shorter string and falling back to ICU
to deal with the remainder.

As a special case for UTF16 Strings, before falling back
to ICU to handle mismatched sub-strings, we take care to
insure that we don't pass sub-strings that begin in the
middle of a collation element.