Treat literal interpolation segments differently
Currently, Swift treats all segments of an interpolation—whether literal or interpolated value—the same: they are each individually passed to `init<T>(stringInterpolationSegment: T)`, then the results are collectively passed to `init(stringInterpolation: Self…)`. So `”hello \(1) world”` becomes:

  .init(stringInterpolation:
    .init(stringInterpolationSegment: “hello “),
    .init(stringInterpolationSegment: 1),
    .init(stringInterpolationSegment: “ world”)
  )

This change does *not* wrap the literal segments in a call to `init(stringInterpolationSegment:)`, so it instead generates:

  .init(stringInterpolation:
    “hello “,
    .init(stringInterpolationSegment: 1),
    “ world”
  )

That has two desirable effects:

1. Only the interpolated values pass through `init(stringInterpolationSegment:)`, so the conforming type can apply special behavior to the interpolated values.

2. Putting the literals in the context of an `init(stringInterpolation:)` call means that they are forced to be `Self` literals, rather than being in the unconstrained generic context of an `init(stringInterpolationSegment:)` call, which ends up creating `String` literals.

This code does not actually work properly yet—it causes an assertion to fail when you use a custom string interpolation—but it’s a good starting point.