SIL: Fix recent bug introduced into old style mangling of specializations
The format is "type conformances... '_'" for each substitutable
generic parameter. I broke it so that it emits types for the
substitutable parameters but conformances for all requirements
in the signature.

This changed the mangling of specializations where the function
being specialized placed additional constraints on associated
types, causing demangling failure in the pre-specialization
whitelist code.

This might fix <rdar://problem/30932656>.