Allow "redeclaring" an initializer with non-overlapping availability.
This allows a library to "fix" an initializer that should have been
failable or non-failable by gating the change on the user's version of
Swift. Failability (and whether or not the initializer throws) is
something that doesn't change the "overload signature", so normally
the compiler would call this an invalid redeclaration. But since only
one version can be active at a time based on the client context, there
won't be a problem in practice.

https://bugs.swift.org/browse/SR-4171 / rdar://problem/30470854

I realized in the process of doing this that methods may also have
mismatched 'throws', and we should allow that as well. Next commit.