[test] Expand collection combinatorics to include index
This adds a custom index to the variations in collection types in
the collection-combinatorics test. This is important because Int has
properties that make it more powerful as an index, particularly for
random-access collections.

Adding that custom index identified a weird edge-case in macOS for a
non-mutable range replaceable random-access collection with a custom
index where the `DefaultRandomAccessIndices` type doesn’t attach
without a type alias. Linux shows the same behavior with a custom-
indexed random-access collection.