Invalidate default argument exprs if typechecking fails
If typechecking fails, the expression will have unsolved type variables
written into it.  This Crashes The Compiler.

In that case, there’s no reason to keep a tree of dangling references
around.  Detach the initializer expression from the AST, but continue
to typecheck it to see if we can get some useful diagnostics out of it.