Sema: Fix assertion when performing conformance check of protocol in another file
GenericSignature::getConformanceAccessPath() would call
computeRequirementSignature() if the protocol did not already
have one.

If this happened before the protocol was validated with
validateDecl(), then we would hit an assertion because
validateDecl() calls computeRequirementSignature()
unconditionally.

Change validateDecl() to only call computeRequirementSignature()
if the protocol does not already have a requirement signature.

Progress on <rdar://problem/30817732>.