[silgen] Add a doxygen doc above FormalEvaluationScope.
I added a general description, but more importantly, I added a note about a
quirk in our current implementation that existed before my current changes.

Specifically:

All formal access contain a pointer to a cleanup in the normal cleanup
stack. This is to ensure that when SILGen calls Cleanups.emitBranchAndCleanups
(and other special cleanup code along error edges), writebacks are properly
created. What is key to notice is that all of these cleanup emission types are
non-destructive. Contrast this with normal scope popping. In such a case, the
scope pop is destructive. This means that any pointers from the formal access to
the cleanup stack is now invalid.

rdar://30955427