[Bridging PCH] Install ASTReaderCallbacks correctly, rdar://31044067.
Previously we installed our ASTReaderListener in clang's CompilerInstance
ModuleManager, in all cases. We use this to track dependencies, so it's
important that it gets installed in the ASTReader that's reading PCH
files, to track the input files that the PCH depends on.

Unfortunately clang's internal handling of -import-pch involves building
a separate, unrelated ASTReader and _not_ propagating ASTReaderHandlers
from the instance. So we've been emitting broken dependency information.

Clang _does_ propagate the ASTDeserializationListener to the new
ASTReader, but that listener only has events that happen too late in
the ASTReader's lifecycle (after it has read the PCH file's control
block, where the list of input files we're interested in is stored), so we
add a hook (in clang) to the ASTDeserializationListener, to install
the ASTReaderListener when we're attached to an ASTReader.